Adds fallback if the dialog fragment doesn't have a dialog.
According to the android docs, DialogFragments are intended to be **optionally** dialogs.  While they can be displayed as a dialog, they can also be displayed as a regular fragment, so it's easy to reuse code in different configurations.

This change accounts for that by falling back to resolving the target view from the fragment's view, in cases where the dialog doesn't exist.